requirement.txt
streamlit
langchain
boto3
python-dotenv
sqlite3
sqlite-utils
langchain-community
langchain-aws
sqlalchemy
chromadb
sentence-transformers
transformers